pyproject.toml
[tool.poetry]
name = "tibber-task"
version = "0.1.0"
description = ""
authors = ["private <oskar.elvkull@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
alembic = "^1.13.3"
uvicorn = "^0.31.0"
pydantic-settings = "^2.5.2"
black = "^24.8.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"
greenlet = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_default_fixture_loop_scope = "function"